// src/api/generated/models/listMyCasesRequestStatus.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ListMyCasesRequestStatus = typeof ListMyCasesRequestStatus[keyof typeof ListMyCasesRequestStatus];


export const ListMyCasesRequestStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  WITHDRAWN: 'WITHDRAWN',
} as const;