// src/api/generated/models/listMyCasesRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ListMyCasesRequestStatus } from './listMyCasesRequestStatus';

export interface ListMyCasesRequest {
  /** @minimum 1 */
  pageNum?: number;
  /** @minimum 1 */
  pageSize?: number;
  status?: ListMyCasesRequestStatus;
}