// src/api/generated/models/pageResultAppVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AppVO } from './appVO';

export interface PageResultAppVO {
  list?: AppVO[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}